meta-aspeed-sdk: Add python3-pyhsslms recipe
The fmc-imgtool uses this pyhsslms python module, adds this recipe to build it
into OpenEmbedded/Yocto build environment.

Currently, the meta-arm has this recipe and it is placed in
meta-arm/meta-arm/recipes-devtools/python/python3-pyhsslms_2.0.0.bb.
However, AST2500 did not add meta-arm layer, so copy this recipe from meta-arm
to meta-aspeed-sdk. It is a temporary solution.

Will upstream this recipe to either OpenEmbedded-Core or met-openembedded

Signed-off-by: Jamin Lin <jamin_lin@aspeedtech.com>
Change-Id: Ib8862a7848e9bcaed21ce83c0d473827d7c52fbe