meta-aspeed-sdk: rename the mdio-tool recipe to pievo-mdio-tool
There is an "mdio-tool" recipe which is placed in "meta-aspeed-sdk" meta layer.
This recipe is used to build the mdio-tool, with its source repository hosted at:
https://github.com/PieVo/mdio-tool

There is another recipe named "mdio-tools" in the "meta-openembedded/meta-networking"
meta layer. This recipe is used to build the mdio-tools, with its source repository
hosted at: https://github.com/wkz/mdio-tools

The names of both recipes are very similar which may confuse users, rename the
mdio-tool recipe to pievo-mdio-tool.

Signed-off-by: Jamin Lin <jamin_lin@aspeedtech.com>
Change-Id: I272e5313155c4032f96bc9f020b210b351f2c4b8